requirements.txt
Flask
mysql-connector-python
werkzeug
pyotp
qrcode
Pillow
